# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: 0.0.1
...
paths:
...
  /users/{UserId}:
    parameters:
      - $ref: "#/components/parameters/UserId"

    put:
      tags: ["user"]
      summary: "The user puts their username"
      description: |-
        The client can change the username of their profile
      operationId: setMyUserName
      requestBody:
        description: |- 
          The username the user wants to change their name to. Must be between 3 and 16 characters
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserName"
              
        required: true
        
      responses:
        '204':
          $ref: "#/components/responses/NoContent"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"

      security:
          - BearerAuth: []
    
    get:
      tags: ["user"]
      summary: "The profile of the user is shown"
      description: |-
        The profile must have all the following attributes: the number of followers, the stream of photos, and on each photo, the comments and the number of likes
      operationId: getUserProfile
      responses:
        '200':
          $ref: "#/components/responses/Profile"
        '206':
          $ref: "#/components/responses/PartialContent"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"
                    
      security:
        - BearerAuth: []
  /users:
    put:
      tags: ["user"]
      summary: "The username of a user is searched"
      description: |-
        In order for the user to search and follow, the username is searched
      operationId: searchUsername
      responses:
        '200':
          $ref: "#/components/responses/Profile"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"
                    
      security:
        - BearerAuth: []
...
  /users/{UserId}/homescreen:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: ["stream"]
      summary: "All the photos of the user are shown"
      description: |-
        The pictures the user has already uploaded should all appear. Also, with each one of them the number of likes it has received and the array of comments it has should appear
      operationId: getMyStream
      responses:
        '200':
          $ref: "#/components/responses/Stream"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"

      security:
        - BearerAuth: []
  /users/{UserId}/photos:
    parameters:
      - $ref: '#/components/parameters/UserId'
    post:
      tags: ["photos"]
      summary: "The user uploads a photo to their profile"
      description: |-
        The client uploads a photo that will appear in their profile 
      operationId: uploadPhoto
      requestBody:
        description: photo
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
        required: true
      responses:
        '201':
          $ref: "#/components/responses/PhotoUploaded"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"

      security:
        - BearerAuth: []
  /users/{UserId}/photos/{PhotoId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/PhotoId'
    
    delete: 
      tags: ["photos"]
      summary: "The user deletes a photo of their profile, and all the associated comments"
      description: |-
        The user deletes a photo that must be in their profile, so both the photo and its associated comments should be deleted
        
      operationId: deletePhoto
        
      responses:
        '204':
          $ref: "#/components/responses/NoContent"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"
                    
      security:
        - BearerAuth: []
    
    get:
      tags: ["photos"]
      summary: "gets a photo to make changes to it"
      description: |-
        The info of a photo is taken to interact with it
      
      operationId: getPhoto
        
      responses:
        '200':
          $ref: "#/components/schemas/Photo"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/ServerError"
      
      security:
        - BearerAuth: []
  /users/{UserId}/photos/{PhotoId}/file:
  
    parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/PhotoId'
  
    get:
        tags: ["photos"]
        summary: "gets the file of the photo"
        description: |-
          The file of the photo is taken to be able to print it on the screen
        
        operationId: getFile
          
        responses:
          '200':
            $ref: "#/components/responses/File"
          '401':
            $ref: "#/components/responses/Unauthorized"
          '403':
            $ref: '#/components/responses/Forbidden'
          '404':
            $ref: "#/components/responses/NotFound"
          '500':
            $ref: "#/components/responses/ServerError"
        
        security:
          - BearerAuth: []
...
components:
  parameters:
    UserId: 
      name: UserId
      in: path
      description: the user id, defined by a number (pk)
      required: true
      schema:
        description: it is an integer between 0-99999
        type: integer
        minimum: 0
        maximum: 99999
        example: 12345
    PhotoId:
      name: PhotoId
      in: path
      description: the photo id, defined by a number (pk)
      required: true
      schema:
        description: integer between 0-99999
        type: integer
        minimum: 0
        maximum: 99999
        example: 13673
...
  schemas:
    UserName:
      title: UserName
      description: the user's username
      type: object
      properties:
        UserName:
          description: the string that will be the username, between 3-16 characters
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: margrey
    User:
      title: User
      description:  The user id and their username
      type: object
      properties:
        UserID: 
          description: the user id, integer from 0-99999
          type: integer
          minimum: 0
          maximum: 99999
          example: 89898
        Username:
          $ref: "#/components/schemas/UserName"
    PhotoId:
      title: PhotoId
      description: the ID of the photo
      type: object
      properties:
        PhotoID: 
          description: the photo id
          type: integer
          minimum: 0
          maximum: 99999
          example: 7483
        UserId:
          description: the user id, integer from 0-99999
          type: integer
          minimum: 0
          maximum: 99999
          example: 897
    Photo:
      title: Photo
      description: the photo schema, including the id, user, likes, date, comments
      type: object
      properties:
        PhotoID:
          $ref: "#/components/schemas/PhotoId"
        UserID:
          description: this is the user that uplodaded the photo id
          type: integer
          minimum: 0
          maximum: 99999
          example: 84938
        UserName:
          $ref: "#/components/schemas/UserName"
        File: 
          description: the file where the photos are stored
          type: string
          format: binary
          minLength: 1
          maxLength: 99999
        Date:
          description: the moment when the photo was uploaded (date and time)
          type: string
          format: date-time
          example: "2023-10-25T14:06:45Z"
          minLength: 3
          maxLength: 99999
        Likes:
          description: array of likes of the photo (users that liked it)
          type: array
          minItems: 0
          maxItems: 99999
          items: 
            $ref: "#/components/schemas/Like"
          example:
            - UserID: 64733
              UserName: margrey
            - UserID: 73829
              UserName: empan
        Comments:
          description: comments of the photo
          type: array
          minItems: 0
          maxItems: 99999
          items:
            $ref: "#/components/schemas/Comment"
          example: 
            - comments:
              - CommentID: 63745
                UserID: 64733
                UserName: margrey
                PhotoID: 78392
                Comment: "Nice picture"
    CommentText:
      title: CommentText
      type: object
      description: the conctent of the comment in a string
      properties:
        CommentText:
          description: the comment that the user posts content
          type: string
          pattern: '^.*?$'
          minLength: 1
          maxLength: 50
          example: That is so coollll
    Comment:
      title: Comment
      type: object
      description: schema of a comment, with the content of it, the user and the photo it belongs to
      properties:
        CommentID: 
          description: the comment id
          type: integer
          minimum: 0
          maximum: 99999
          example: 7383
        UserID:
          description: the user that posts the comment id
          type: integer
          minimum: 0
          maximum: 99999
          example: 78987
        UserName:
          $ref: "#/components/schemas/UserName"
        PhotoID:
          $ref: "#/components/schemas/PhotoId"
        Comment:
          $ref: "#/components/schemas/CommentText"
      example:
        CommentID: 56746
        UserID: 74833
        UserName: margrey
        PhotoID: 83938
        Comment: That is so coollll
    Like:
      title: Like
      type: object
      description: a like, with its id, the photo and the person who likes
      properties:
        LikeID:
          description: the unique like id
          type: integer
          minimum: 0
          maximum: 99999
          example: 665
        UserID:
          description: the user that likes the photo id
          type: integer
          minimum: 0
          maximum: 99999
          example: 637
        PhotoID:
          $ref: "#/components/schemas/PhotoId"
      example:
        LikeID: 738
        UserID: 839
        PhotoID: 83938
    Profile:
      description: contains all the info of a profile, followers, following, photos with all their info
      type: object
      properties:
        UserID:
          description: the user that posts the comment id
          type: integer
          minimum: 0
          maximum: 99999
          example: 78987
        UserName:
          $ref: "#/components/schemas/UserName"
        Followers:
          description: the array of followers of the user
          type: array
          minItems: 0
          maxItems: 99
          items: 
            $ref: "#/components/schemas/User"
        Following:
          description: the array of users the user follows
          type: array
          minItems: 0
          maxItems: 99
          items: 
            $ref: "#/components/schemas/User"
        Photos:
          description: the array of photos the user has posted
          type: array
          minItems: 0
          maxItems: 99
          items: 
            $ref: "#/components/schemas/Photo"
          example:
            - Comments:
                  - UserID: 8987
                    UserName: "empan"
                    CommentID: 637
                    Comment: "Nice picture"
                    
                  - user_id: 73829
                    nickname: "haruno"
                    CommentID: 7383
                    comment: "that is great"
                    
            - likes:
              - UserID: 8987
                UserName: "empan"
              - UserID: 73829
                UserName: "haruno"
                
              date: 2023-09-21T09:12:56Z
              PhotoID: 74839
              owner: "margrey"
  responses:
    NoContent:
      description: "The response for code 200, where there is no content"
    Stream:
      description: Contains a list of photos from following users in a reversed chronological order
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Photo"
...
    PhotoUploaded:
      description: 'the photo was added successfully, so the photo is returned'
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Photo"
    Profile:
      description: All the info of a profile
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Profile"
    PartialContent:
      description: "206 - partial content, useful for banned users"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Profile"
    File:
      description: "File of a photo"
      content:
        multipart/form-data:
          schema:
            type: object
            properties:
              photo:
                type: string
                format: binary
                description: "The photo file"
    Unauthorized:
      description: "Error 401 - Unauthorized - The user is not logged in"
    Forbidden:
      description: "Error 403 - Forbidden - That action cannot be executed"
    NotFound:
      description: "Error 404 - Not found"
    ServerError:
      description: "Error 500 - Internal server error occurred"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer